# social_graph_service/src/config/swaggerDoc.yaml
openapi: 3.0.1
info:
  title: moody Social Graph Service
  description: This is the REST API documentation for the moody social_graph_service
  version: 1.0.0
servers:
  - url: /api
...
paths:
  /profiles:
    post:
      tags:
        - profile
      summary: Create a new profile or update an existing one
      description: If the user (provided via the header) does not have a profile yet, a new profile will be created. Else the profile will be updated.
      requestBody:
        description: Profile object to be created or updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        201:
          description: Profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  profile:
                    $ref: '#/components/schemas/Profile'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    get:
      tags:
        - profile
      summary: Get all profiles
      description: Fetching all exisiting profiles.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  profiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
        404:
          description: No profiles found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No profiles found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
    delete:
      tags:
        - profile
      summary: Delete all registered profiles
      description: all existing profiles are deleted (for testing purpose only)
      responses:
        200:
          description: All profiles removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  msg:
                    type: string
                    example: All profiles removed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
  /profiles/me:
    get:
      tags:
        - profile
      summary: Get my profile
      description: through the auth header, the user object will be submitted and the current user's profile will be returned
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  profile:
                    type: object
                    $ref: '#/components/schemas/Profile'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No profile found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No profile found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
  /profiles/{username}:
    get:
      tags:
        - profile
      summary: Get profile by username
      description: If there is a profile for the username, this profile object will be returned
      parameters:
        - in: path
          name: username
          schema:
            type: string
            example: Hugo
          required: true
          description: the username of a profile
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  profile:
                    type: object
                    $ref: '#/components/schemas/Profile'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No profile found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No profile found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
...
  /messages:
    post:
      tags:
        - message
      summary: Create a new message
      description: A new message is created and linked to a profile.
      requestBody:
        description: The message I want to send.
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: My personal message :)
                emoji:
                  type: string
                  example: '🤢'
        required: true
      responses:
        201:
          description: Profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  message:
                    $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    get:
      tags:
        - message
      summary: Get all messages
      description: Fetching all existing messages.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        404:
          description: No profiles found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    delete:
      tags:
        - message
      summary: Delete all messages
      description: all existing messages are deleted (for testing purpose only)
      responses:
        200:
          description: All profiles removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  msg:
                    type: string
                    example: All messages removed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
...
  /search/messages:
    get:
      tags:
        - search
      summary: Get messages from a specific search term
      description: Search for messages of my followed profiles given a specific search term (text)
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
            example: bla
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No messages found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
components:
  schemas:
    Profile:
      required:
        - birthday
        - hometown
        - gender
      type: object
      properties:
        birthday:
          type: string
          format: date
        hometown:
          type: string
          example: Linz
        gender:
          type: string
          example: female
          enum:
            - male
            - female
            - transgender
        privacy:
          type: boolean
          example: false
          default: false
        notifications:
          type: boolean
          example: false
          default: false
        username:
          type: string
          example: Hugo
    Message:
      required:
        - text
        - emoji
      type: object
      properties:
        text:
          type: string
          example: My first message
        emoji:
          type: string
          example: 🤢
        author:
          type: string
          example: Hugo
        timestamp:
          type: string
          format: date-time
          example: Thu Dec 03 2020 13:23:42 GMT+0000 (Coordinated Universal Time)
  securitySchemes:
    authentication:
      type: apiKey
      name: Authorization
      in: header